fixed the file structure inside of the post dir
