+++ auth/routes/__test__/signin.test.js
@@ -1,9 +1,9 @@
 const request = require('supertest');
-const app = require('../../../src/app');
+const app = require('../../../app');
 
 it('fails when a email that does not exist is supplied', async () => {
 	await request(app)
-		.post('/api/signin')
+		.post('/users/api/signin')
 		.send({
 			email: 'test@test.com',
 			name: 'name',
@@ -15,7 +15,7 @@ it('fails when a email that does not exist is supplied', async () => {
 it('fails the signin when a wrong password is provided', async () => {
 	// send a post request to signup a user
 	await request(app)
-		.post('/api/signup')
+		.post('/users/api/signup')
 		.send({
 			email: 'test@test.com',
 			name: 'name',
@@ -26,7 +26,7 @@ it('fails the signin when a wrong password is provided', async () => {
 
 	// send a post request to signin with the wrong password
 	await request(app)
-		.post('/api/signup')
+		.post('/users/api/signup')
 		.send({
 			email: 'test@test.com',
 			password: 'wrong',
@@ -38,7 +38,7 @@ it('fails the signin when a wrong password is provided', async () => {
 it('responds with a cookie when given valid credentials', async () => {
 	// send a post request to signup a user
 	await request(app)
-		.post('/api/signup')
+		.post('/users/api/signup')
 		.send({
 			email: 'test@test.com',
 			name: 'name',
@@ -49,7 +49,7 @@ it('responds with a cookie when given valid credentials', async () => {
 
 	// send a post request to signin with valid credentials
 	await request(app)
-		.post('/api/signin')
+		.post('/users/api/signin')
 		.send({
 			email: 'test@test.com',
 			password: 'password',
