+++ auth/routes/__test__/signout.test.js
@@ -1,9 +1,9 @@
 const request = require('supertest');
-const app = require('../../app');
+const app = require('../../../app');
 
 it('clears the cookie after signing out', async () => {
 	await request(app)
-		.post('/api/signup')
+		.post('/users/api/signup')
 		.send({
 			email: 'test@test.com',
 			name: 'name',
@@ -11,7 +11,10 @@ it('clears the cookie after signing out', async () => {
 		})
 		.expect(201);
 
-	const res = await request(app).post('/api/signout').send({}).expect(200);
+	const res = await request(app)
+		.post('/users/api/signout')
+		.send({})
+		.expect(200);
 
 	expect(res.get('Set-Cookie')).not.toBeDefined();
 });
