+++ package.json
@@ -10,7 +10,7 @@
   "jest": {
     "testEnvironment": "node",
     "setupFilesAfterEnv": [
-      "./src/test/setup.js"
+      "./test/setup.js"
     ]
   },
   "keywords": [],
@@ -18,6 +18,7 @@
   "license": "MIT",
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "body-parser": "^1.19.0",
     "cookie-session": "^1.4.0",
     "express": "^4.17.1",
     "express-validator": "^6.10.0",
